apply clippy suggestions

+++ crates/wasmi/src/table/mod.rs
@@ -506,7 +506,7 @@ impl TableEntity {
             .and_then(|items| items.get(..len_size))
             .ok_or(TrapCode::TableOutOfBounds)?;
         if let Some(fuel) = fuel {
-            fuel.consume_fuel_if(|costs| costs.fuel_for_copies(u64::from(len)))?;
+            fuel.consume_fuel_if(|costs| costs.fuel_for_copies(len))?;
         }
         // Finally, copy elements in-place for the table.
         dst_items.copy_from_slice(src_items);
@@ -542,7 +542,7 @@ impl TableEntity {
             return Err(TrapCode::TableOutOfBounds);
         };
         if let Some(fuel) = fuel {
-            fuel.consume_fuel_if(|costs| costs.fuel_for_copies(u64::from(len)))?;
+            fuel.consume_fuel_if(|costs| costs.fuel_for_copies(len))?;
         }
         // Finally, copy elements in-place for the table.
         self.elements
